fix: PR updates and fixes
