+++ src/state/hooks.ts
@@ -546,22 +546,16 @@ export const useLiveIfoStatus = () => {
   return { liveIfos: LiveIfo }
 }
 
-// TAGS
-export const useFetchLiveTags = () => {
+export const useFetchLiveTagsAndOrdering = () => {
   const dispatch = useAppDispatch()
   const { slowRefresh } = useRefresh()
   useEffect(() => {
+    dispatch(fetchLiveOrdering())
     dispatch(fetchLiveTags())
   }, [dispatch, slowRefresh])
 }
 
-export const useFarmTags = (chainId: number) => {
-  const { Tags }: StatsState = useSelector((state: State) => state.stats)
-  const farmTags = Tags?.[`${chainId}`].farms
-
-  return { farmTags }
-}
-
+// TAGS
 export const usePoolTags = (chainId: number) => {
   const { Tags }: StatsState = useSelector((state: State) => state.stats)
   const poolTags = Tags?.[`${chainId}`]?.pools
@@ -569,39 +563,10 @@ export const usePoolTags = (chainId: number) => {
   return { poolTags }
 }
 
-export const useJungleFarmTags = (chainId: number) => {
-  const { Tags }: StatsState = useSelector((state: State) => state.stats)
-  const jungleFarmTags = Tags?.[`${chainId}`]?.jungleFarms
-
-  return { jungleFarmTags }
-}
-
 // ORDERING
-export const useFetchLiveOrdering = () => {
-  const dispatch = useAppDispatch()
-  const { slowRefresh } = useRefresh()
-  useEffect(() => {
-    dispatch(fetchLiveOrdering())
-  }, [dispatch, slowRefresh])
-}
-
-export const useFarmOrderings = (chainId: number) => {
-  const { Ordering }: StatsState = useSelector((state: State) => state.stats)
-  const farmOrderings = Ordering?.[`${chainId}`]?.farms
-
-  return { farmOrderings }
-}
-
 export const usePoolOrderings = (chainId: number) => {
   const { Ordering }: StatsState = useSelector((state: State) => state.stats)
   const poolOrderings = Ordering?.[`${chainId}`]?.pools
 
   return { poolOrderings }
-}
-
-export const useJungleFarmOrderings = (chainId: number) => {
-  const { Ordering }: StatsState = useSelector((state: State) => state.stats)
-  const jungleFarmOrderings = Ordering?.[`${chainId}`]?.jungleFarms
-
-  return { jungleFarmOrderings }
 }
