+++ src/views/DualFarms/DualFarms.tsx
@@ -2,9 +2,10 @@ import React, { useState, useRef, useEffect } from 'react'
 import { useLocation } from 'react-router-dom'
 import BigNumber from 'bignumber.js'
 import { Flex } from '@apeswapfinance/uikit'
-import { useFarmOrderings, useFarmTags, useFetchFarmLpAprs } from 'state/hooks'
+import { useFetchFarmLpAprs } from 'state/hooks'
 import { useDualFarms, usePollDualFarms } from 'state/dualFarms/hooks'
-import { DualFarm, Farm } from 'state/types'
+import { useFarmOrderings, useFarmTags } from 'state/farms/hooks'
+import { DualFarm } from 'state/types'
 import { orderBy } from 'lodash'
 import ListViewLayout from 'components/layout/ListViewLayout'
 import Banner from 'components/Banner'
@@ -119,11 +120,13 @@ const DualFarms: React.FC = () => {
 
     switch (sortOption) {
       case 'all':
-        return orderBy(
-          farms,
-          (farm: DualFarm) => farmOrderings.find((ordering) => ordering.pid === farm.pid)?.order,
-          'asc',
-        ).slice(0, numberOfFarmsVisible)
+        return farmOrderings
+          ? orderBy(
+              farms,
+              (farm: DualFarm) => farmOrderings.find((ordering) => ordering.pid === farm.pid)?.order,
+              'asc',
+            ).slice(0, numberOfFarmsVisible)
+          : farms.slice(0, numberOfFarmsVisible)
       case 'stables':
         return farms
           .filter(
@@ -146,11 +149,13 @@ const DualFarms: React.FC = () => {
       case 'liquidity':
         return orderBy(farms, (farm: DualFarm) => parseFloat(farm.totalStaked), 'desc').slice(0, numberOfFarmsVisible)
       default:
-        return orderBy(
-          farms,
-          (farm: DualFarm) => farmOrderings.find((ordering) => ordering.pid === farm.pid)?.order,
-          'asc',
-        ).slice(0, numberOfFarmsVisible)
+        return farmOrderings
+          ? orderBy(
+              farms,
+              (farm: DualFarm) => farmOrderings.find((ordering) => ordering.pid === farm.pid)?.order,
+              'asc',
+            ).slice(0, numberOfFarmsVisible)
+          : farms.slice(0, numberOfFarmsVisible)
     }
   }
