+++ src/views/Farms/Farms.tsx
@@ -3,14 +3,14 @@ import { useLocation } from 'react-router-dom'
 import BigNumber from 'bignumber.js'
 import useActiveWeb3React from 'hooks/useActiveWeb3React'
 import { Flex } from '@apeswapfinance/uikit'
-import { useFarmOrderings, useFarmTags, useFetchFarmLpAprs, useFetchLpTokenPrices } from 'state/hooks'
+import { useFetchFarmLpAprs, useFetchLpTokenPrices } from 'state/hooks'
 import ListViewMenu from 'components/ListViewMenu'
 import { orderBy } from 'lodash'
 import ListViewLayout from 'components/layout/ListViewLayout'
 import Banner from 'components/Banner'
 import { useTranslation } from 'contexts/Localization'
 import { Farm } from 'state/types'
-import { useFarms, usePollFarms } from 'state/farms/hooks'
+import { useFarms, usePollFarms, useFarmOrderings, useFarmTags } from 'state/farms/hooks'
 import DisplayFarms from './components/DisplayFarms'
 import { BLUE_CHIPS, NUMBER_OF_FARMS_VISIBLE, STABLES } from './constants'
 import HarvestAllAction from './components/CardActions/HarvestAllAction'
@@ -108,11 +108,13 @@ const Farms: React.FC = () => {
 
     switch (sortOption) {
       case 'all':
-        return orderBy(
-          farms,
-          (farm: Farm) => farmOrderings.find((ordering) => ordering.pid === farm.pid)?.order,
-          'asc',
-        ).slice(0, numberOfFarmsVisible)
+        return farmOrderings
+          ? orderBy(
+              farms,
+              (farm: Farm) => farmOrderings.find((ordering) => ordering.pid === farm.pid)?.order,
+              'asc',
+            ).slice(0, numberOfFarmsVisible)
+          : farms.slice(0, numberOfFarmsVisible)
       case 'stables':
         return farms
           .filter((farm) => STABLES.includes(farm.tokenSymbol) && STABLES.includes(farm.quoteTokenSymbol))
@@ -138,11 +140,13 @@ const Farms: React.FC = () => {
           'asc',
         ).slice(0, numberOfFarmsVisible)
       default:
-        return orderBy(
-          farms,
-          (farm: Farm) => farmOrderings.find((ordering) => ordering.pid === farm.pid)?.order,
-          'asc',
-        ).slice(0, numberOfFarmsVisible)
+        return farmOrderings
+          ? orderBy(
+              farms,
+              (farm: Farm) => farmOrderings.find((ordering) => ordering.pid === farm.pid)?.order,
+              'asc',
+            ).slice(0, numberOfFarmsVisible)
+          : farms.slice(0, numberOfFarmsVisible)
     }
   }
